- introduced operator<< for enums - fixed minor issues

+++ srg_worldmodel/include/srg/dialogue/Task.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <srgsim/SRGEnums.h>
-#include <srgsim/containers/Coordinate.h>
 #include <essentials/IdentifierConstPtr.h>
+#include <srgsim/containers/Coordinate.h>
+#include <srgsim/world/TaskType.h>
 namespace srg
 {
 class SRGWorldModel;
@@ -18,7 +18,7 @@ public:
 
     friend std::ostream& operator<<(std::ostream& os, const srg::dialogue::Task& obj)
     {
-        os << "Task: Type: " << static_cast<int>(obj.type) << " at " << obj.coordinate;
+        os << "Task: Type: " << obj.type << " at " << obj.coordinate;
         return os;
     }
